+++ QMUIKit/UIKitExtensions/QMUICollectionViewPagingLayout.m
@@ -181,7 +181,7 @@ const CGFloat QMUICollectionViewPagingLayoutRotationRadiusAutomatic = -1.0;
         }
     }
     
-    proposedContentOffset.x = roundf(proposedContentOffset.x / itemSpacing) * itemSpacing;
+    proposedContentOffset.x = round(proposedContentOffset.x / itemSpacing) * itemSpacing;
     
     return proposedContentOffset;
 }
